requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
argcomplete==1.10.0
attrs==24.2.0
auto-py-to-exe==2.44.2
bcrypt==4.3.0
beautifulsoup4==4.8.0
blinker==1.8.2
bottle==0.13.2
bottle-websocket==0.2.9
cachetools==5.5.0
camelot-py==0.11.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clipboard==0.0.4
clr-loader==0.2.6
colorama==0.4.6
colour==0.1.5
compressed_rtf==1.0.6
comtypes==1.4.7
contourpy==1.3.1
cryptography==43.0.3
cssselect2==0.7.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
distro==1.9.0
docx2pdf==0.1.8
docx2txt==0.8
ebcdic==1.1.1
EbookLib==0.17.1
Eel==0.17.0
et_xmlfile==2.0.0
extract-msg==0.23.1
fake-useragent==2.0.3
fastapi==0.116.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.55.3
frozenlist==1.5.0
future==1.0.0
fuzzywuzzy==0.18.0
gevent==24.10.3
gevent-websocket==0.10.1
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.66.2
gspread==6.1.3
gTTS==2.5.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
IMAPClient==2.1.0
immutabledict==4.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keyboard==0.13.5
kiwisolver==1.4.7
latex2mathml==3.77.0
Levenshtein==0.26.1
lxml==5.2.2
Markdown==3.7
MarkupSafe==3.0.1
matplotlib==3.10.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
ntplib==0.4.0
Nuitka==2.4.10
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.46
openai==1.51.2
opencv-python==4.10.0.84
openpyxl==3.1.5
ordered-set==4.1.0
ortools==9.9.3963
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
panflute==2.3.1
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pefile==2024.8.26
pillow==11.1.0
propcache==0.3.0
proto-plus==1.24.0
protobuf==5.26.1
proxy_tools==0.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pyfiglet==1.0.2
pygame==2.6.1
PyGetWindow==0.0.9
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyLaTeX==1.4.2
pylatexenc==2.10
PyMsgBox==1.0.9
PyMuPDF==1.24.11
pypandoc==1.15
pyparsing==3.1.4
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
pypiwin32==223
PyQt5-Qt5==5.15.2
PyQt6-WebEngine==6.7.0
PyQt6-WebEngine-Qt6==6.7.3
PyQt6-WebEngineSubwheel-Qt6==6.7.3
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
PyRect==0.2.0
PySCIPOpt==5.5.0
PyScreeze==1.0.1
PySide6==6.8.0.1
PySide6_Addons==6.8.0.1
PySide6_Essentials==6.8.0.1
PySocks==1.7.1
python-constraint==1.4.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-Levenshtein==0.26.1
python-pptx==0.6.18
pythonnet==3.0.4
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.2
pywebview==5.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.1
QtAwesome==1.4.0
QtPy==2.4.3
RapidFuzz==3.10.1
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9
scipy==1.14.1
selenium==4.25.0
setuptools==75.1.0
shiboken6==6.8.0.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
SpeechRecognition==3.8.1
SQLAlchemy==2.0.38
starlette==0.47.3
styleframe==4.2
svglib==1.5.1
sympy==1.13.3
tabula-py==2.10.0
tabulate==0.9.0
textract==1.6.3
thefuzz==0.22.1
tikzplotlib==0.10.1
tinycss2==1.4.0
tk==0.1.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.4.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.0.4
whichcraft==0.6.1
WMI==1.5.1
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.2.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.1.0
zstandard==0.23.0
